protein_agent/requirements.txt
openai>=1.65,<2
fastapi>=0.110
uvicorn>=0.30
MDAnalysis>=2.7.0
biopython>=1.83
numpy>=1.26
pyyaml>=6.0
rich>=13.0
typer>=0.12
loguru>=0.7
jsonschema>=4.22
